+++ tests/test_parallel.rs
@@ -22,7 +22,7 @@ fn parallel_count_single_file() {
     let wrk = Workdir::new("parallel_count_single_file");
 
     let mut cmd = wrk.command("parallel");
-    cmd.arg("count").arg("-F").arg(wrk.resource("series.csv"));
+    cmd.arg("count").arg(wrk.resource("series.csv"));
 
     let got: String = wrk.stdout(&mut cmd);
 
@@ -36,7 +36,6 @@ fn parallel_count_single_file_preprocess() {
     let mut cmd = wrk.command("parallel");
     cmd.arg("count")
         .args(["-P", "search -es Category Disc"])
-        .arg("-F")
         .arg(wrk.resource("series.csv"));
 
     let got: String = wrk.stdout(&mut cmd);
@@ -110,7 +109,6 @@ fn parallel_freq_single_file() {
     let mut cmd = wrk.command("parallel");
     cmd.arg("freq")
         .args(["-s", "Category"])
-        .arg("-F")
         .arg(wrk.resource("series.csv"));
 
     let got: Vec<Vec<String>> = wrk.read_stdout(&mut cmd);